requirements.txt
attrs==20.3.0
flake8==3.9.0
iniconfig==1.1.1
mccabe==0.6.1
packaging==20.9
paho-mqtt==1.5.1
pluggy==0.13.1
py==1.10.0
pycodestyle==2.7.0
pyflakes==2.3.1
pyparsing==2.4.7
pytest==6.2.2
toml==0.10.2
yapf==0.31.0
